fix LIVECD user id conflict between casper and virtualbox
Virtualbox creates a user with uid 999.
Casper's LIVECD user is also hard-coded to use uid 999.  The
resulting conflict prevents login from the live cd.

So, if you are creating your master image in VirtualBox and are
using the guest extensions, you will hit this conflict.
The workaround tries to use uid 990, and provides an error
message if both uids are unavailable.