init.d/firstboot: cleanup and add teamviewer config removal
Errors are output to /dev/null anyway, and no big deal if there
is a missing package.  It just makes it look ugly and error-prone
to check for the packages existence first since there is no "clean"
way to check if a package is installed.

Relocated a teamview .conf removal introduced in PinguyBuilder to
firstboot cleanup.