fix undefined variable  DISTRIB_RELEASE
The LiveCDLabel is used for too many purposes.
-casper desktop installer icon
	-which reads /cdrom/.disk/info
-grub boot menu
-cdrom label

casper desktop - uses $1 $2 for the installer name, and replaces any
dashes with a space.  Updated .conf defaults to encourage this.
In practise, it doesn't appear as if the industry too stringently
follows any conventions for .disk/info

Assume that $2 can act as a version number.  No harm done if
the convention is not properly followed.

casper replaces dashes with a space.