boot-repair needs a PPA - adding documentation code
If boot-repair is available somehow, it will be installed.
However, that is not common.  Yet, we don't want to prompt
the user in the middle of remastersys for anything, so
we can't easily add a PPA.

Not sure how to best handle this, so the code to add the PPA
is all there, but commented out.  In our case, we use
wasta-offline, so boot-repair is always available...