add ability to specify slideshow in .conf
If memory serves correctly, the dependency on ubiquity-slideshow
was simply to avoid a large-ish download at the time of running
remastersys. However, it tended to just download random versions,
contrary to the intention.

In any case, if we allow the user to specify the slideshow, then
nothing should be "pre-downloaded" anyway.

Allow the user to avoid slideshows all together by leaving that
entry blank.

The old pre/postinst routines weren't updated on an earlier
.conf location change. Fix that, and ensure that on upgrading,
an unsuspecting user isn't left without a slideshow.