etc/wasta-remastersys/pool: allow pre-downloaded .deb files
requested by Asad: for EFI files in case of network troubles.
This is generic enough so that any deb files can be provided for
offline installation if desired.

This is pretty hidden away - you have to review the source code
to know that it is possible. I'm not really sure IF/HOW this
can be used by a user interactively, so there doesn't seem
to be any need to "advertise" it.