Eoan support: to enable ZFS in the installer, Recommend: zfsutils-linux
... and zfs-initramfs.

New in 19.10 is experimental support for zfs on root.
This install option is hidden if zfs packages are not installed.

Adding this as a recommends ought to satisfy the intention
of ensuring that it is installed (since Ubuntu normally
installs recommended packages), but it won't block the
installation if it is not available.

It seemed like a slightly cleaner solution than adding
another apt-get install line in remastersys, but mostly
6 of 1, half a dozen of the other.

Even better would be if Ubuntu added this recommends
to ubiquity, but their discussion suggests they want
each distribution to "seed" this.