make slideshow install 3x quieter
Especially if you have run remastersys several times,
there will be some "orphaned" packages from the ubquity removal.
Since ubiquity is about to be installed again, there is
no need to see all that information.

If the preferred slideshow fails to install,
we DO want the operator to know about that,
since it is a user defined setting,
so don't send everything to /dev/null.

Three --quiets seems to do the trick.