Eoan suport: try to handle patches better
Patches are going to be for specific versions of software,
so how best to handle a patch in a multi-version .deb package?

We have two examples here - one where the 18.04 patch now seems
unnecessary, and one where the patch is still valuable 2 years later.

My solution was to test if the patch applies without error - in
that case apply and expect that it will work properly in the future.
If the patch fails, then notify about the failure, but skip applying
portions of the patch (since in this example one hunk succeeded and
one hunk failed).