4.5.7: 20.04 - don't set root to 777 (since Calamares doesn't fix)
mksquashfs is not told to include root, it is only
given a list of subfolders. So it created the root
node with 777 permission. Other installers obviously
must "fix" this up, but Calamares doesn't.

VERY recently (only in 20.04), mksquashfs added
an option to set the -root-mode to whatever
permission you want. Since the other work-arounds
didn't work for me, and since we only expect
to use Calamares installer for 20.04 or later,
I just decided to use that new feature, and
ignore the problem for older distros.

Thanks to ufw status (firewall) for alerting me to
the problematic 777 mode for /.

Calamares was notified of the problem
(https://github.com/calamares/calamares/issues/1418)
but didn't really seem to care.