4.5.8: also cleanup linux-modules to avoid adding an unsigned kernel
The internet talks about 4.15 linux-modules for 18.04, so perhaps
this has been around longer than I thought.

The problem appears to be that we remove linux-image and linux-header,
but leave behind linux-modules - which depends on either linux-image
or linux-image-unsigned. Since we are removing the linux-image,
the left-behind linux-modules fulfills its dependency
by installing the -unsigned kernel.

Well, the intention is to fully remove anything about
older kernels, so adding linux-modules to the list
should take care of the problem.