make wasta-remastersys calamares-aware
Calamares is a generic Linux installer.
It is used by Lubuntu and StudioLinux
within the Ubuntu family.
(Also Manjaro, KDE neon, OpenMandriva...)

With Ubuntu potentially dropping support for ubiquity,
(since it is developing an alternative in flutter),
it would be nice to support Calamares well.

One nice thing - it even works (as a live CD)
when you pick "just install" from the initial boot menu.
It fails to find the ubiquity installer,
and just boots up into live mode anyway.