warn if genisoimage and snaps have not been removed.
The normal "mkisofs" provided on Ubuntu is genisoimage,
which has a limit of a 4GB file able to be added to an ISO.
Since the file system is squashed into one big file,
this limit is easily reached if snaps are included.

Unfortunately the alternative mkisofs is not easy to find,
so it is hard to recommend something to the user
at this point, other than to remove snaps.

wasta-core provides wasta-remove-snaps, which keeps
apt snapd installed, but which removes all of the snaps
(which would be updated immediately after OS install anyway).

In my testing, the entire Ubuntu 24.04 (without snaps)
took 2GB, but with snaps it exceeded the limit
(with basically only firefox as a functional snap).

Without the warning, the user would go through
the entire ISO build process, only to fail at the last step.