mark cryptsetup/lvm2 as manually installed, avoiding orphaning
If you installed with encrypting the disk,
then it worked on first boot, but not on subsequent boots.

The problem is that ubiquity pulled these in as dependencies,
and when the installer was removed they were later autoremoved.

It is NOT enough to just put this into the firstboot service,
in case something goes wrong there. (I just put it here as
a double safety precaution.)